% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anntpkglist}
\alias{anntpkglist}
\title{Genome and Annotaion packages in Bioconductor}
\format{
\subsection{\code{anntpkglist} A dataframe with 32 rows and three columns;}{

rows describe different species
\describe{
\item{genome}{Name of the genome package}
\item{species}{Scientific names of species}
\item{annotation}{Name of the annotation species}
...
}
}
}
\source{
\url{https://www.bioconductor.org/packages/release/data/annotation/}
}
\usage{
anntpkglist
}
\description{
Chart for latest genome and annotation package for few selected species that
are downloaded when no reference is provided by user
}
\keyword{datasets}
