% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createdir.R
\name{createdir}
\alias{createdir}
\title{create the directories for saving the result}
\usage{
createdir(pos = 1, outdir, startover)
}
\arguments{
\item{pos}{enviromental}

\item{outdir}{where the directory and its subfolder will be created}

\item{startover}{: if the run is again?}
}
\value{
list of folder names
}
\description{
create the directories for saving the result
}
