% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadcompoundset.R
\name{loadcsets}
\alias{loadcsets}
\title{Function to generate compound set for pathway analysis}
\usage{
loadcsets(entitycode)
}
\arguments{
\item{entitycode}{three char code of organism for kegg}
}
\value{
compound_sets is returned
}
\description{
This function list the pathways along with the list of compounds all the
pathway contains
}
