% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{onexistcleanup}
\alias{onexistcleanup}
\title{Function to clean the reference directories/packages}
\usage{
onexistcleanup(ref.dir, entity)
}
\arguments{
\item{ref.dir}{: directory for reference files}
}
\description{
Function to clean the reference directories/packages
}
