% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genedata.R
\name{plot_genedata}
\alias{plot_genedata}
\title{Plot the genes belonging to top significant pathways
whose gene expression is more than that of background genes}
\usage{
plot_genedata(
  gene_data,
  gage.out,
  gset,
  outdir,
  mref,
  msamp,
  compare,
  q_cutoff
)
}
\arguments{
\item{gene_data}{obj returned by differential analysis function}

\item{gage.out}{result of gage analysis for logFC data}

\item{gset}{gene set of interest}

\item{outdir}{directory to store the heatmaps}

\item{mref}{reference index for gene/cpd data}

\item{msamp}{sample index for gene/cpd data}

\item{compare}{: GAGE parameter}

\item{q_cutoff}{q val  to select pathways}
}
\value{
nothing returned
}
\description{
\code{plot_genedata} Plot the genes belonging to top significant pathways
whose expression is more than that of background genes for all the samples
genes counts are normalized counts using deseq2
}
