% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_qalign.R
\name{plotalignmentstats}
\alias{plotalignmentstats}
\title{Plot the alignment mapping statistics}
\usage{
plotalignmentstats(data_to_plot, aligned_bam)
}
\arguments{
\item{data_to_plot}{data to plot alignment stats}

\item{aligned_bam}{directory for bam files}
}
\value{
returns after completing plot
}
\description{
Plot the alignment mapping statistics
}
