% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_deseq2.R
\name{plotdeseqheatmap}
\alias{plotdeseqheatmap}
\title{plot the result of result of deseq2}
\usage{
plotdeseqheatmap(deseq2_res, dds, deseq2.dir, npca, nheatmap)
}
\arguments{
\item{deseq2_res}{result of deseq2}

\item{dds}{deseq2 object}

\item{deseq2.dir}{directory where aligned bam are stored}

\item{npca}{number of genes to use for pca}

\item{nheatmap}{number of gens to be used for heatmap}
}
\value{
vsd_matrix just returns
}
\description{
plot the result of result of deseq2
}
