% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpathways.R
\name{plotpathways}
\alias{plotpathways}
\title{Check and plot pathways with the data}
\usage{
plotpathways(
  kegg.dir,
  entity,
  path.ids,
  gene_data,
  cpd_data,
  cpd_id_type = "kegg",
  gene_id_type = "entrez"
)
}
\arguments{
\item{kegg.dir}{directory in which the pathways are saved}

\item{entity}{organism whose pathway is plotted}

\item{path.ids}{list of pathway ids to plot c("00056","00098")}

\item{gene_data}{gene information to be plotted in pathview}

\item{cpd_data}{compound data to plot}

\item{cpd_id_type}{what type of cpd.id is used in cpd.data}

\item{gene_id_type}{what type of gene id is used in gene_data}
}
\value{
invisiblenull
}
\description{
This function checks if pathways can be downloaded from the kegg and
plots by rendering the data
}
