% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_phenofile.R
\name{process_phenofile}
\alias{process_phenofile}
\title{process phenofile
this function extracts information from user supplied phenofile ;
information include samplename, location of raw files, if the reads are paired,
which samples belong to which class and which sample is pair of other sample}
\usage{
process_phenofile(phenofile)
}
\arguments{
\item{phenofile}{path to the phenofile where raw data files path is stored}
}
\value{
a list of things
}
\description{
process phenofile
this function extracts information from user supplied phenofile ;
information include samplename, location of raw files, if the reads are paired,
which samples belong to which class and which sample is pair of other sample
}
