% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combinedpath_analysis.R
\name{run_combinedpath_analysis}
\alias{run_combinedpath_analysis}
\title{Run the gene and compound set analysis
runs gage for gene and compound data and combines the result using gamma function}
\usage{
run_combinedpath_analysis(
  outdir,
  gene_data,
  entity,
  gcompare,
  gage.out.cpd_res,
  qcut
)
}
\arguments{
\item{outdir}{main directory for storing output of the process}

\item{gene_data}{logfold change data or cnts}

\item{entity}{Scientific name of species whose RNA/compound is being analyzed}

\item{gcompare}{how samples should be compared for gene data}

\item{gage.out.cpd_res}{the out object of cpound analysis}

\item{qcut}{threshold for pathway selection}
}
\value{
pathids important pathways of interest
}
\description{
Run the gene and compound set analysis
runs gage for gene and compound data and combines the result using gamma function
}
