% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cpathway_analysis.R
\name{run_compound_kegg_gsets}
\alias{run_compound_kegg_gsets}
\title{Function for running the compound pathway analysis}
\usage{
run_compound_kegg_gsets(
  cdatapath,
  cpd_id_type = "KEGG COMPOUND accession",
  csamp,
  keggorgcode,
  cref,
  ccompare = "paired",
  outdir,
  q_cutoff
)
}
\arguments{
\item{cdatapath}{:data path for compound data}

\item{cpd_id_type}{type of compound id as compatible to ones in
pathview::data(rn.list)}

\item{csamp}{column number of sample}

\item{keggorgcode}{three char code for kegg organism}

\item{cref}{column number of reference}

\item{ccompare}{comparison type for sample}

\item{outdir}{directory where results of compound set analysis is stored}

\item{q_cutoff}{q value cutoff for selecting pathway}
}
\value{
gage_return_obj_c return object from run_gage for compound
}
\description{
This function prepares for running the gage analysis on compound sets
}
