% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_deseq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run standard DESeq2 for differential gene expression analysis and plot
volcano plots}
\usage{
run_deseq2(cnts, outdir, entity, formula_object, coldat, npca, nheatmap)
}
\arguments{
\item{cnts}{:counts of gene as data.frame or filename of count of gene,one count per sample per gene , sample in row}

\item{outdir}{: directory to store results of deseq2}

\item{entity}{: scientific name of organism to convert gene to symbol}

\item{formula_object}{string denoting formula}

\item{coldat}{the dataframe for phenotype information of all sample}

\item{npca}{number of genes to use for pca}

\item{nheatmap}{number of genes for heatmap}
}
\value{
fold change values
}
\description{
run_deseq2 takes counts and the list indicating reference and samples and the
directory where the results are stored and performs the deseq2 analysis
The output is result table with columns of genes and log2FoldChange from
result of deseq2 analysis and a volcanoplot.
For a paired experiment, sample name of the pair should be same but their class is different
It returns the log2foldChange to be used by GAGE for gene set analysis.
}
