% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_go_gsets.R
\name{run_go_set_analysis}
\alias{run_go_set_analysis}
\title{run go enrichmnet analysis}
\usage{
run_go_set_analysis(
  gene_data,
  cnts,
  outdir,
  entity,
  compare,
  use.fold,
  phenofile
)
}
\arguments{
\item{gene_data}{logfold change, names should be genes}

\item{cnts}{gene counts}

\item{outdir}{main directory for storing output of the process}

\item{entity}{Scientific name of species whose RNA is being analyzed}

\item{compare}{how the comparision is made in gage}

\item{use.fold}{if fold change should be used instead}

\item{phenofile}{file where the path of raw data is stored, used to generate ref and samp index}
}
\description{
This function runs go enrichment analysis for biological process, cellular component and metabolic function separately
}
