% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kegg_gset_analysis.R
\name{run_kegg_analysis}
\alias{run_kegg_analysis}
\title{run KEGG enrichmnet analysis}
\usage{
run_kegg_analysis(gene_data, outdir, entity, compare)
}
\arguments{
\item{gene_data}{dataframe or matrix or list of genes}

\item{outdir}{directory to store output}

\item{entity}{organism of interest}

\item{compare}{how the comparision is made}
}
\description{
This function runs go enrichment analysis for KEGG disease, metabolism and signalling.
An additional analysis is also done by combining kegg metabolism and signalling pathways
}
