% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_qCount.R
\name{run_qCount}
\alias{run_qCount}
\title{Count the number of genes expression and store count in RDS file}
\usage{
run_qCount(aligned_proj_lst, corenum, outdir, entity, references)
}
\arguments{
\item{aligned_proj_lst}{:the name of R object saved or generated after all alignments}

\item{corenum}{: the number of cores used for alignment}

\item{outdir}{: the directory in which result is stored}

\item{entity}{: the scientific name of the species of interest}

\item{references}{: the references which is used for alignment and counting}
}
\value{
count of genes
}
\description{
generate the count of gene and if gene id is ensembl, it converts it to
entrez to match with genes names for gene set analysis.
the generated table with genes in row and counts in columns are stored as a
RDS file that can be loaded into R for future analysis.
}
