% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanity_check}
\alias{sanity_check}
\title{function to make sure things are good}
\usage{
sanity_check(ref.dir, outdir, pos = 1, entity, corenum, compare, rerun)
}
\arguments{
\item{ref.dir}{: directory for reference files}

\item{outdir}{: directory for result}

\item{pos}{: environment specifier}

\item{entity}{: scientific name of species of interest}

\item{corenum}{: number of cores}

\item{compare}{: comparision to make for gage}

\item{rerun}{if FALSE the previously complete step will not be rerunned, if
TRUE analysis starts from first step}
}
\value{
list of different paths for result files
}
\description{
function to make sure things are good
}
