% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writesampleFile.R
\name{writesampleFile}
\alias{writesampleFile}
\title{Write sampleFile to run qAlign}
\usage{
writesampleFile(outdir, phenofile_res, nchunks)
}
\arguments{
\item{outdir}{path to result dir where the file is written}

\item{phenofile_res}{: object of process phenofile}

\item{nchunks}{how many pieces should the phenofile be cut to}
}
\value{
sampleFile sampleFile path to be used by qAlign
}
\description{
This function creates the sampleFile to run qAlign
}
