% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bods.R
\docType{data}
\name{bods}
\alias{bods}
\title{Dataset from gage package}
\format{
\subsection{\code{bods}}{

A dataframe with x rows and y columns;
\describe{
\item{genome}{Name of the genome package}
\item{species}{Scientific names of species}
\item{annotation}{Name of the annotation species}
...
}
}
}
\source{
gage
}
\usage{
data(bods)
}
\value{
the data frame GAGE with information about species
}
\description{
Chart for latest genome and annotation package for few selected species that
are downloaded when no reference is provided by user
}
\keyword{datasets}
