% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_references}
\alias{check_references}
\title{function to make sure things are good}
\usage{
check_references(ref.dir, entity)
}
\arguments{
\item{ref.dir}{: directory for reference files}

\item{entity}{: scientific name of species of interest}
}
\value{
list of different paths for result files
}
\description{
function to make sure things are good
}
