% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createdir.R
\name{createdir}
\alias{createdir}
\title{create the directories for saving the result}
\usage{
createdir(pos = 1, outdir, entity, startover, keep_tmp, is.test)
}
\arguments{
\item{pos}{enviromental}

\item{outdir}{where the directory and its subfolder will be created}

\item{entity}{: scientific name of the organism}

\item{startover}{: if the run is again?}

\item{keep_tmp}{: keep bam files resulting from alignment or not}

\item{is.test}{to ask for prompt during test , is.test = TRUE, for test}
}
\value{
list of folder names
}
\description{
create the directories for saving the result
}
