% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareNalignNcount.R
\name{do_alignment_ncounting}
\alias{do_alignment_ncounting}
\title{Align and count
This function identifies the type of data(SE or PE), creates sampleFile neccessary for qAlign,
checks if references are okay for alignement and counting, runs qAlign and qcount}
\usage{
do_alignment_ncounting(
  phenofile,
  ref.dir,
  outdir,
  entity,
  cacheDir,
  aligner,
  corenum,
  nchunks
)
}
\arguments{
\item{phenofile}{: path to the phenofile where raw data files path is stored}

\item{ref.dir}{directory where reference genome and annotation is stored}

\item{outdir}{directory for output files}

\item{entity}{organism of interest}

\item{cacheDir}{: directory where temporary files generated during alignment}

\item{aligner}{: weather Rhisat2 or Rbowtie should be used for alignment}

\item{corenum}{number of cores to run alignment}

\item{nchunks}{if the alignment should be done in chunks}
}
\value{
lists with names cnts and aligned_proj
}
\description{
Align and count
This function identifies the type of data(SE or PE), creates sampleFile neccessary for qAlign,
checks if references are okay for alignement and counting, runs qAlign and qcount
}
