% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwrap.R
\name{pathwrap}
\alias{pathwrap}
\title{Wrapper for RNASeq data analysis from raw reads to pathway visualization}
\usage{
pathwrap(
  phenofile,
  entity,
  corenum = detectCores(),
  ref.dir = NULL,
  cacheDir = tempdir(),
  outdir = NULL,
  startover = FALSE,
  aligner = "Rhisat2",
  diff.tool = "DESeq2",
  cnts = NA,
  gcompare = "unpaired",
  npca = 19,
  nheatmap = 10,
  fc_matrix = NA,
  cdatapath = NA,
  cpd_id_type = "KEGG COMPOUND accession",
  csamp = NULL,
  cref = NULL,
  ccompare = "unpaired",
  qcut = 0.01,
  pathids = "04110",
  nchunks = 1,
  keep_tmp = TRUE,
  cpd.idtype = "KEGG"
)
}
\arguments{
\item{phenofile}{file where the path of raw data is stored.}

\item{entity}{Scientific name of species whose RNA is being analyzed}

\item{corenum}{number of cores avaialble for run}

\item{ref.dir}{path to reference directory which contain
reference file(\emph{.fa) and annotation file(}.gtf), can be NA}

\item{cacheDir}{directory where temporary files created during alignment}

\item{outdir}{main directory for storing output of the process}

\item{startover}{do you want to start from beginning, default =False}

\item{aligner}{One of "Rhisat2" or "Rbowtie2"; Rbowtie2 can be very slow
for human and eukaryotic species}

\item{diff.tool}{weather to use "DESeq2" or edgeR for differential gene analysis}

\item{cnts}{file path to the cnts per gene from other analysis}

\item{gcompare}{how the comparision is done for transcripts/genes}

\item{npca}{number of genes to use for pca}

\item{nheatmap}{number of genes for heatmap}

\item{fc_matrix}{logfold change for gage if deseq2 is not run}

\item{cdatapath}{data path for compound data}

\item{cpd_id_type}{"KEGG COMPOUND accession"}

\item{csamp}{index/row number where sample files are ex: c(5,6,4)}

\item{cref}{index/row number where references are, ex: c(1,2,3)}

\item{ccompare}{how the compound data is compared, default paired}

\item{qcut}{threshold for pathway selection, default 0.01}

\item{pathids}{pathway of interest only necessary if enrichment is not run}

\item{nchunks}{default 1, in how many chunks you want to run alignment}

\item{keep_tmp}{weather to store aligned bam files and trimmed fastq files default True}
}
\value{
returns message if analysis complete successfully
}
\description{
wrapper that does quality control analysis of raw files, performs adapter and
quality trimming, builds genome index and does alignment , counts genes
perfoms differential gene expression analysis,does gene enrichment test using
GAGE and visualize the enriched pathways using pathview all using one
wrapper function. It has the ability to continue the analysis if it is
halted at any stage and generate quality pictures and generate comprehensive
analysis of the data.
}
