% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotpathways.R
\name{plotpathways}
\alias{plotpathways}
\title{Check and plot pathways with the data}
\usage{
plotpathways(gage.dir, entity, path.ids, fc_matrix, cpd_data = NULL)
}
\arguments{
\item{gage.dir}{directory in which the pathways are saved}

\item{entity}{organism whose pathway is plotted}

\item{path.ids}{list of pathway ids to plot}

\item{fc_matrix}{genedata to plot}

\item{cpd_data}{compound data to plot}
}
\value{
invisiblenull
}
\description{
This function checks if pathways can be downloaded from the kegg and
plots by rendering the data
}
