% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_analysis.R
\name{run_combinedpath_analysis}
\alias{run_combinedpath_analysis}
\title{script to run the combine analysis of compound and gene
This is internal function of pathviewwrap}
\usage{
run_combinedpath_analysis(
  gpath_ids,
  cpath_ids,
  gsets,
  pgs.gene,
  pgs.cpd,
  somdir,
  gage.out,
  gage.out.cpd,
  qcut = 0.01
)
}
\arguments{
\item{gpath_ids}{top pathways from gene based analysis}

\item{cpath_ids}{significance compound sets ids}

\item{gsets}{pathways on which analysis is based}

\item{pgs.gene}{p-value of analysis of gene set}

\item{pgs.cpd}{p-value of analysis of compound set}

\item{somdir}{directory where data is stored}

\item{gage.out}{retsult from gage analysis of gene}

\item{gage.out.cpd}{retsult from gage analysis of compound}

\item{qcut}{the threshold value to determine significance of gene/cpd sets}
}
\value{
path.ids significant pathids from combined analysis
}
\description{
script to run the combine analysis of compound and gene
This is internal function of pathviewwrap
}
