% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_deseq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Title}
\usage{
run_deseq2(cnts, grp.idx, deseq2.dir)
}
\arguments{
\item{cnts}{:counts of gene}

\item{grp.idx}{: index of reference and samples for differential analysis}

\item{deseq2.dir}{: directory to store results of deseq2}
}
\value{
exp.fc : fold change values
}
\description{
Title
}
