% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_deseq2.R
\name{run_deseq2}
\alias{run_deseq2}
\title{Run standard DESeq2 for differential gene expression analysis and plot
volcano plots}
\usage{
run_deseq2(cnts, grp.idx, deseq2.dir, entity, SampleName = NULL, compare)
}
\arguments{
\item{cnts}{:counts of gene}

\item{grp.idx}{: index of reference and samples for differential analysis}

\item{deseq2.dir}{: directory to store results of deseq2}

\item{entity}{: scientific name of organism to convert gene to symbol}

\item{SampleName}{list of the sample name to be used in paired experiment}

\item{compare}{: if the experiment is paired or unpaired}
}
\value{
fold change values
}
\description{
run_deseq2 takes counts and the list indicating reference and samples and the
directory where the results are stored and performs the deseq2 analysis
The output is result table with columns of genes and log2FoldChange from
result of deseq2 analysis and a volcanoplot.
It returns the log2foldChange to be used by GAGE for gene set analysis.
}
