% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_differerntial_gene_analysis.R
\name{run_differerntial_gene_analysis}
\alias{run_differerntial_gene_analysis}
\title{Prepare to run standard DESeq2 or edgeR for differential gene expression analysis and plot
volcano plots}
\usage{
run_differerntial_gene_analysis(
  cnts,
  phenofile_res,
  outdir,
  entity,
  gcompare,
  npca,
  nheatmap,
  diff.tool = "DeSeq2"
)
}
\arguments{
\item{cnts}{:counts of gene as data.frame or filename of count of gene,one count per sample per gene , sample in row}

\item{phenofile_res}{returned object from process_phenofile}

\item{outdir}{: directory to store results of deseq2}

\item{entity}{: scientific name of organism to convert gene to symbol}

\item{gcompare}{: if the experiment is paired or unpaired}

\item{npca}{number of genes to use for pca}

\item{nheatmap}{number of genes for heatmap}

\item{diff.tool}{either DeSeq2 or edgeR}
}
\value{
fold change values
}
\description{
Prepare to run standard DESeq2 or edgeR for differential gene expression analysis and plot
volcano plots
}
