% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_edgeR.R
\name{run_edgeR}
\alias{run_edgeR}
\title{Title}
\usage{
run_edgeR(cnts, grp.idx, edger.dir)
}
\arguments{
\item{cnts}{: counts of genes}

\item{grp.idx}{: index of the reference and sample for differential analysis}

\item{edgeR.dir}{: the directory in which edgeR results will be stored}
}
\value{
exp.fc : log fold expression values
}
\description{
Title
}
