% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_fastp.R
\name{run_fastp}
\alias{run_fastp}
\title{Runs fastp for quality and adapter trimming}
\usage{
run_fastp(sampleName)
}
\arguments{
\item{sampleName}{: name of the sample}
}
\value{
no value returned
}
\description{
This function takes name of the samples and for each sample does the quality
and adapter trimming for Illumina and long read sequencing. It works for both
PE and SE data
}
