% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_GAGE.R
\name{run_gage}
\alias{run_gage}
\title{Run GAGE and Pathview}
\usage{
run_gage(gsets, work.dir, same.dir, compare, fc_matrix, entity)
}
\arguments{
\item{gsets}{: gene sets to analyse}

\item{work.dir}{: directory where results will be stored}

\item{same.dir}{: if the direction is same for GAGE analysis, GAGE parameter}

\item{compare}{: GAGE parameter}

\item{fc_matrix}{: log fold change in matrix form}

\item{entity}{scientific name of organism}
}
\value{
nothing returned
}
\description{
\code{run_pathway} runs GAGE for pathway analysis
GAGE is based upon the user supplied comparision
method for the species specified.
The biological process, cellular component and
molecular function analysis for GO terms are done separately.
}
\details{
KEGG disease, KEGG signalling and metabolism pathways are analysed separately
Top enriched pathways with  "q.val" < 0.1 are visualized using pathview.
}
