% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gage.R
\name{run_gage2}
\alias{run_gage2}
\title{Run GAGE and Pathview}
\usage{
run_gage2(
  gene_data,
  gsets,
  same.dir,
  compare,
  work.dir,
  gene_id_type = "ENTREZ",
  ref = NULL,
  samp = NULL
)
}
\arguments{
\item{gene_data}{gene data to run gene set analysis when available}

\item{gsets}{: gene sets to analyse}

\item{same.dir}{: if the direction is same for GAGE analysis, GAGE parameter}

\item{compare}{: GAGE parameter}

\item{work.dir}{: directory where results will be stored}

\item{gene_id_type}{what type of id is gene_id_type in}

\item{ref}{reference index for gene/cpd data}

\item{samp}{sample index for gene/cpd data}
}
\value{
nothing returned
}
\description{
\code{run_pathway} runs GAGE for pathway analysis
GAGE is based upon the user supplied comparision
method for the species specified.
The biological process, cellular component and
molecular function analysis for GO terms are done separately.
}
\details{
KEGG disease, KEGG signalling and metabolism pathways are analysed separately
Top enriched pathways with  "q.val" < 0.1 are visualized using pathview.
}
