% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_gset_analysis.R
\name{run_gene_gsets_analysis}
\alias{run_gene_gsets_analysis}
\title{Run GAGE and Pathview}
\usage{
run_gene_gsets_analysis(
  gene_data,
  outdir,
  entity,
  compare,
  phenofile_res,
  q_cutoff
)
}
\arguments{
\item{gene_data}{: gene data with gene_data$logfoldchange #log fold expression values}

\item{outdir}{: directory in which GAGE will be run}

\item{entity}{organism of interest, scientific name}

\item{compare}{: how the comparison is done for GAGE, see gage for details}

\item{phenofile_res}{: object returned by process_phenofile, see process phenofile for details}

\item{q_cutoff}{q val for pathway selection}
}
\value{
gage_return_obj gage_return_obj which is list is returned
}
\description{
\code{run_pathway} runs GAGE for pathway analysis
GAGE is based upon the user supplied comparison
method for the species specified.
The biological process, cellular component and
molecular function analysis for GO terms are done separately.
}
\details{
KEGG disease, KEGG signalling and metabolism pathways are analysed separately
Top enriched pathways with  "q.val" < 0.1 are visualized using pathview.
}
