% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_gset_analysis.R
\name{run_gene_gsets_analysis}
\alias{run_gene_gsets_analysis}
\title{Run GAGE and Pathview}
\usage{
run_gene_gsets_analysis(fc_matrix, cnts, outdir, entity, compare, phenofile)
}
\arguments{
\item{fc_matrix}{: log fold expression values}

\item{cnts}{: counts of genes to use in pathview}

\item{outdir}{: directory in which GAGE will be run}

\item{entity}{organism of interest, scientific name}

\item{compare}{: how the comparison is done for GAGE, see gage for details}

\item{phenofile}{: path to the phenofile where raw data files path is stored}
}
\value{
gage_return_obj gage_return_obj which is list is returned
}
\description{
\code{run_pathway} runs GAGE for pathway analysis
GAGE is based upon the user supplied comparision
method for the species specified.
The biological process, cellular component and
molecular function analysis for GO terms are done separately.
}
\details{
KEGG disease, KEGG signalling and metabolism pathways are analysed separately
Top enriched pathways with  "q.val" < 0.1 are visualized using pathview.
}
