% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_go_gsets.R
\name{run_go_set_analysis}
\alias{run_go_set_analysis}
\title{run go enrichmnet analysis}
\usage{
run_go_set_analysis(
  gene_data,
  outdir,
  entity,
  compare,
  use.fold = T,
  mref,
  msamp,
  q_cutoff
)
}
\arguments{
\item{gene_data}{object returned from prev names should be genes}

\item{outdir}{main directory for storing output of the process}

\item{entity}{Scientific name of species whose RNA is being analyzed}

\item{compare}{how the comparision is made in gage}

\item{use.fold}{if fold change should be used instead}

\item{mref}{reference  colums for plotting gene data}

\item{msamp}{sample columns for plotting gene data}

\item{q_cutoff}{q value for pathway selection}
}
\description{
This function runs go enrichment analysis for biological process, cellular component and metabolic function separately
}
