% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_gset_analysis.R
\name{run_pathway}
\alias{run_pathway}
\title{Run GAGE and Pathview}
\usage{
run_pathway(entity, exp.fc, compare, gage.dir, cnts)
}
\arguments{
\item{entity}{: scientific name of the species}

\item{exp.fc}{: log fold expression values}

\item{compare}{: how the comparison is done for GAGE, see gage for details}

\item{gage.dir}{: directory in which GAGE will be run}

\item{cnts}{: counts of genes to use in pathview}
}
\value{
gage_return_obj gage_return_obj which is list is returned
}
\description{
\code{run_pathway} runs GAGE for pathway analysis
GAGE is based upon the user supplied comparision
method for the species specified.
The biological process, cellular component and
molecular function analysis for GO terms are done separately.
}
\details{
KEGG disease, KEGG signalling and metabolism pathways are analysed separately
Top enriched pathways with  "q.val" < 0.1 are visualized using pathview.
}
