% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_qalign.R
\name{run_qAlign}
\alias{run_qAlign}
\title{\enumerate{
\item RUN THE ANALYSIS # Alignment and counting
}}
\usage{
run_qAlign(phenofile, cacheDir, aligner, references, outdir, corenum, nchunks)
}
\arguments{
\item{phenofile}{: path to the phenofile where raw data files path is stored}

\item{cacheDir}{: directory where temporary files generated during alignment
are store}

\item{aligner}{: weather Rhisat2 or Rbowtie should be used for alignment}

\item{references}{: the list of references genome file and gene ananotaion used for
alignment or bioconductor genome package}

\item{outdir}{directory where aligned bam files are stored}

\item{corenum}{: the number of cores used during alignment}

\item{nchunks}{the number of segment you want to run the alignemnt}
}
\value{
R object generated from the alignment step
}
\description{
this function runs the Rhisat2 or Rbowtie for alignment on paired or
single end mode. It saves the alignment object in RDS file which can
be loaded in R for further analysis. If the reference index is not found
in the reference directory, it creates reference index before running
alignment. If the reference genome is a package the reference index is
created as R package.It generates the barplot for number of mapped and
unmapped sequence reads.
}
