% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_fastqc.R
\name{run_qc}
\alias{run_qc}
\title{Title}
\usage{
run_qc(fq.dir, qc.dir, corenum)
}
\arguments{
\item{fq.dir}{: the directory in which raw RNAseq files are stored}

\item{qc.dir}{: the directory in which results of quality check are stored}

\item{corenum}{: the number of cores used for running quality check}
}
\description{
Title
}
