% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_qCount.R
\name{run_qCount}
\alias{run_qCount}
\title{Title}
\usage{
run_qCount(aligned_proj, corenum, result.dir, txdb, entity)
}
\arguments{
\item{aligned_proj}{: the name of R object saved or generated during alignment}

\item{corenum}{: the number of cores used for alignment}

\item{result.dir}{: the directory in which result is stored}

\item{txdb}{: the txdb object which is used for counting}

\item{entity}{: the scientific name of the species of interest}
}
\value{
cnts : count of genes
}
\description{
Title
}
