% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_rna_qCount.R
\name{run_qCount}
\alias{run_qCount}
\title{Count the number of genes expression and store count in RDS file}
\usage{
run_qCount(aligned_proj, corenum, result.dir, txdb, entity)
}
\arguments{
\item{aligned_proj}{: the name of R object saved or generated during alignment}

\item{corenum}{: the number of cores used for alignment}

\item{result.dir}{: the directory in which result is stored}

\item{txdb}{: the txdb object which is used for counting}

\item{entity}{: the scientific name of the species of interest}
}
\value{
cnts : count of genes
}
\description{
generate the count of gene and if gene id is ensembl, it converts it to entrez to match with genes names for gene set analysis.
the generated table with genes in row and counts in columns are stored as a RDS file that can be loaded into R for future analysis.
}
