% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writesampleFile.R
\name{writesampleFile}
\alias{writesampleFile}
\title{Write sampleFile to run qAlign}
\usage{
writesampleFile(outdir, filenames, SampleName)
}
\arguments{
\item{outdir}{:  path to result dir where the file is written}

\item{filenames}{: list of filenames extracted from phenofile}
}
\value{
statement about the function run
}
\description{
This function creates the sampleFile to run qAlign
}
